requirements.txt

fastapi==0.114.2
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.36
asyncpg==0.29.0
pydantic==2.9.2
python-dotenv==1.0.1
psycopg[binary]>=3.2
python-dotenv>=1.0
tabulate>=0.9
